feat: 유저 관련이 아닌 불필요한 equals hashcode 삭제 및 Password equals 추가
