+++ src/main/java/shopping/domain/Product.java
@@ -1,6 +1,5 @@
 package shopping.domain;
 
-import java.util.Objects;
 import javax.persistence.AttributeOverride;
 import javax.persistence.AttributeOverrides;
 import javax.persistence.Column;
@@ -58,24 +57,5 @@ public class Product {
 
     public Long getPrice() {
         return price.getPrice();
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Product product = (Product) o;
-        return Objects.equals(id, product.id) && Objects.equals(name, product.name)
-                && Objects.equals(imageUrl, product.imageUrl) && Objects.equals(
-                price, product.price);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, name, imageUrl, price);
     }
 }
