+++ src/main/java/shopping/domain/Quantity.java
@@ -1,6 +1,5 @@
 package shopping.domain;
 
-import java.util.Objects;
 import javax.persistence.Embeddable;
 
 @Embeddable
@@ -29,22 +28,5 @@ public class Quantity {
 
     public Integer getQuantity() {
         return quantity;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Quantity quantity1 = (Quantity) o;
-        return Objects.equals(quantity, quantity1.quantity);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(quantity);
     }
 }
